refactor: move Annotation component
